BitCoinStreamData/publisher/requirements.txt
websockets
google-cloud-pubsub
python-dotenv
